+++ tests/testBoard.py
@@ -1,3 +1,4 @@
+"""
 import unittest
 from chess.board import Board
 from chess.rook import Rook
@@ -140,3 +141,5 @@ class TestBoard(unittest.TestCase):
 
 if __name__ == '__main__':
     unittest.main()
+
+"""