+++ tests/testPiezas.py
@@ -1,3 +1,4 @@
+"""
 import unittest
 from chess.pieces import Piece
 from chess.board import Board
@@ -70,3 +71,4 @@ class TestPieces(unittest.TestCase):
 
 if __name__ == '__main__':
     unittest.main()
+"""