+++ tests/testReina.py
@@ -1,3 +1,4 @@
+"""
 import unittest
 from chess.queen import Queen
 from chess.board import Board
@@ -45,3 +46,4 @@ class TestQueenValidPositions(unittest.TestCase):
 
 if __name__ == "__main__":
     unittest.main()
+"""