+++ tests/testRey.py
@@ -1,3 +1,4 @@
+"""
 import unittest
 from chess.king import King
 from chess.board import Board
@@ -83,3 +84,4 @@ class TestKing(unittest.TestCase):
 
 if __name__ == '__main__':
     unittest.main()
+"""